move back all the displaced structs
