+++ send-transaction-service/src/lib.rs
@@ -1,12 +1,6 @@
 #![allow(clippy::arithmetic_side_effects)]
 pub mod send_transaction_service;
-pub mod send_transaction_service_stats;
 pub mod tpu_info;
-pub mod transaction_client;
-pub use {
-    send_transaction_service_stats::SendTransactionServiceStats,
-    transaction_client::{CurrentLeaderInfo, LEADER_INFO_REFRESH_RATE_MS},
-};
 
 #[macro_use]
 extern crate solana_metrics;
